Extract effects into its own header file
